refactor(cqrs): executor changed to message sender
